app-api/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.13.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
